🐳 Feat: image processing functions to use core::Tensor<float32>
- Updated common.cpp to change Tensor<int> to Tensor<float32> for image handling.
- Created a new processing.hpp header file for image processing functionalities.
- Modified transformations.hpp to replace cv::Mat with core::Tensor<float32> in function signatures.
- Implemented changes in transformations.cpp to adapt image processing functions to use core::Tensor<float32>.
- Ensured proper handling of image channels and dimensions across all transformations.
- Added error handling for unsupported channel configurations in various functions.

+++ src/core/include/core/common.hpp
@@ -13,14 +13,17 @@
 #include <iostream>
 #include <opencv2/opencv.hpp>
 
+#include "core.hpp"
+
 namespace core {
 
     std::pair<size_t, size_t> getScreenResolution();
     void showImageCenterWindow(const cv::Mat &image, const std::string &windowName = "Window");
-    void showImageCenterWindow(const Tensor<int> &tensor, const std::string &windowName = "Window");
+    void showImageCenterWindow(const Tensor<float32> &tensor,
+                               const std::string &windowName = "Window");
 
-    void matToTensor(const cv::Mat &src, Tensor<int> &tensor, size_t channels = 3);
-    void tensorToMat(const Tensor<int> &tensor, cv::Mat &dst, size_t channels = 3);
+    void matToTensor(const cv::Mat &src, Tensor<float32> &tensor, int channels = 3);
+    void tensorToMat(const Tensor<float32> &tensor, cv::Mat &dst, int channels = 3);
 }   // namespace core
 
 #endif   // COMMON_HPP
